% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcgamedeals.R
\name{get_active_stores}
\alias{get_active_stores}
\title{Active store IDs and Store names}
\usage{
get_active_stores()
}
\value{
A dataframe with currently active stores and the store IDs being monitored by the API
}
\description{
This function fetches all the active store IDs
and the store names in the API
}
\examples{
get_active_stores()
}
