% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcgamedeals.R
\name{get_deals_by_store_df}
\alias{get_deals_by_store_df}
\title{Get details about a store's maximum deals}
\usage{
get_deals_by_store_df(storeID)
}
\arguments{
\item{storeID}{integer. The store ID mentioned by the user}
}
\value{
A dataframe containing the name of the game,
the store ID, the normal price of a game, the sale price on a game,
the savings and the deal rating off the game.
}
\description{
This function fetches the details
about the given store
}
\examples{
get_deals_by_store_df(2)
}
