% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcgamedeals.R
\name{get_game_reviews_df}
\alias{get_game_reviews_df}
\title{Get details about a store's game reviews}
\usage{
get_game_reviews_df(storeID, rating)
}
\arguments{
\item{storeID}{integer. The store ID mentioned by the user}

\item{rating}{integer. The minimum rating mentioned by the user}
}
\value{
A dataframe containing the name of the game
and the meta critic score of the game.
The meta critic score will be above the one mentioned by the user.
}
\description{
This function fetches the rating for
all the games in a store
}
\examples{
get_game_reviews_df(31,80)
}
